inference/requirements.txt
pillow==11.2.1
pandas==2.3.1
runpod==1.7.13
sqlalchemy==2.0.43
